fix(p5): C2 fractional-weight storage + C3 room_type query correctness
Opus review 2026-04-22 flagged two P5-S1-wiring-time blockers:

C2: alpha/beta/judge_alpha/judge_beta in router_bucket_stats were integer.
Design §6.2 (Oliver Q2) has JUDGE_ALPHA_WEIGHT=0.5 mapping each judge
rating to 0.5 contribution. Integer storage truncates. Changed to
numeric(10,2). Math kernel already handles non-integer alpha/beta.

C3: refresh-router-bucket-stats.ts queried prompt_lab_sessions.room_type,
which doesn't exist in that schema. Actual source is
prompt_lab_iterations.analysis_json->>'room_type'. Without the fix, all
buckets would collapse to 'unknown'. Fixed the SELECT + the row shape.
Falls back to sessions.archetype if analysis_json is absent.

Tests + tsc clean on P5 branch.

+++ supabase/migrations/038_thompson_router.sql
@@ -11,10 +11,13 @@ CREATE TABLE IF NOT EXISTS router_bucket_stats (
   room_type   text      NOT NULL,
   camera_movement text  NOT NULL,
   sku         text      NOT NULL,
-  alpha       integer   NOT NULL DEFAULT 0,
-  beta        integer   NOT NULL DEFAULT 0,
-  judge_alpha integer   NOT NULL DEFAULT 0,
-  judge_beta  integer   NOT NULL DEFAULT 0,
+  -- numeric(10,2) so fractional judge weights (JUDGE_ALPHA_WEIGHT=0.5 per
+  -- Oliver's Q2 decision) persist without truncation. Human ratings still
+  -- contribute integer counts via the same columns; mixing is safe.
+  alpha       numeric(10,2) NOT NULL DEFAULT 0,
+  beta        numeric(10,2) NOT NULL DEFAULT 0,
+  judge_alpha numeric(10,2) NOT NULL DEFAULT 0,
+  judge_beta  numeric(10,2) NOT NULL DEFAULT 0,
   enabled     boolean   NOT NULL DEFAULT true,
   last_updated timestamptz NOT NULL DEFAULT now(),
   UNIQUE (room_type, camera_movement, sku)
